+++ src/gui/skydiveorbust/skydiveorbust_player.h
@@ -10,7 +10,7 @@ int pg_sdob_player_chaptersUpdating;
 void pg_sdob_player_video_chapterMarks(gslc_tsGui *pGui);
 void pg_sdob_player_video_chapterList(int len);
 void pg_sdob_player_video_chapters();
-void pg_sdob_player_sliderTicks(gslc_tsGui *pGui, double *tickMarks, int tickCnt);
+void pg_sdob_player_sliderTicks(gslc_tsGui *pGui, uint16_t * tickMarks, uint8_t tickCnt);
 // void setVideoCurrentTime(long long int video_pos);
 void setSliderPos(gslc_tsGui *pGui, int16_t nPercent);
 void setSliderPosByTime(gslc_tsGui *pGui);
